% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGeoArrowPolygonLayer.R
\name{addGeoArrowPolygonLayer}
\alias{addGeoArrowPolygonLayer}
\title{Add Deck.gl PolygonLayer to a \code{\link[mapgl:maplibre]{mapgl::maplibre()}} or \code{\link[mapgl:mapboxgl]{mapgl::mapboxgl()}} map
using blazing fast \code{\link[nanoarrow:read_nanoarrow]{nanoarrow::write_nanoarrow()}} data transfer.}
\usage{
addGeoArrowPolygonLayer(
  map,
  data,
  layer_id,
  geom_column_name = attr(data, "sf_column"),
  popup = NULL,
  tooltip = NULL,
  render_options = renderOptions(),
  data_accessors = dataAccessors(),
  popup_options = popupOptions(),
  tooltip_options = tooltipOptions(),
  ...
)
}
\arguments{
\item{map}{the \code{\link[mapgl:maplibre]{mapgl::maplibre()}} or \code{\link[mapgl:mapboxgl]{mapgl::mapboxgl()}} map to add the layer to.}

\item{data}{a sf \verb{(MULTI)POLYGON} object.}

\item{layer_id}{the layer id.}

\item{geom_column_name}{the name of the geometry column of the sf object.
It is inferred automatically if only one is present.}

\item{popup}{should a popup be contructed? If \code{TRUE}, will create a popup fromm all
available attributes of the feature. Can also be a character vector of column
names, on which case the popup will include only those columns. If a single character
is supplied, then this will be shown for all features. If \code{NULL} (deafult) or
\code{FALSE}, no popup will be shown.}

\item{tooltip}{should a tooltip be contructed? If \code{TRUE}, will create a tooltip fromm all
available attributes of the feature. Can also be a character vector of column
names, on which case the tooltip will include only those columns. If a single character
is supplied, then this will be shown for all features. If \code{NULL} (deafult) or
\code{FALSE}, no tooltip will be shown.}

\item{render_options}{a list of \link{renderOptions}}

\item{data_accessors}{a list of \link{dataAccessors}}

\item{popup_options}{a list of \link{popupOptions}}

\item{tooltip_options}{a list of \link{tooltipOptions}}

\item{...}{currently not used.}
}
\description{
Add Deck.gl PolygonLayer to a \code{\link[mapgl:maplibre]{mapgl::maplibre()}} or \code{\link[mapgl:mapboxgl]{mapgl::mapboxgl()}} map
using blazing fast \code{\link[nanoarrow:read_nanoarrow]{nanoarrow::write_nanoarrow()}} data transfer.
}
