% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{renderOptions}
\alias{renderOptions}
\title{Deck.gl render options}
\usage{
renderOptions(...)
}
\arguments{
\item{...}{named options to be passed to the relevant deck.gl JavaScript Method.}
}
\value{
List with named options, possibly modified via \code{...} argument.
}
\description{
In deck.gl every layer type has a specific set of render options, see e.g.
those for \href{https://deck.gl/docs/api-reference/layers/scatterplot-layer#render-options}{ScatterPlotLayer}.
This function sets all defaults for all available layer functions in this
package.
Please refer to the relevant \href{https://deck.gl/docs/api-reference/layers/}{deck.gl documentation}
for a more detailed description of the available layer functions.
See Details for a list of currently available options, their defaults and the
layer types they apply to.
}
\details{
Currently, the following options are automatically set to the following defaults:

\strong{ScatterplotLayer}
\itemize{
\item radiusUnits = "pixels"
\item radiusScale = 1
\item lineWidthUnits = "pixels"
\item lineWidthScale = 1
\item stroked = TRUE
\item filled = TRUE
\item radiusMinPixels = 3
\item radiusMaxPixels = 15
\item lineWidthMinPixels = 0
\item lineWidthMaxPixels = 15
\item billboard = FALSE
\item antialiasing = FALSE
}

\strong{PathLayer}
\itemize{
\item widthUnits = "pixels"
\item widthScale = 1
\item widthMinPixels = 1
\item widthMaxPixels = 5
\item capRounded = TRUE
\item jointRounded = FALSE
\item miterLimit = 4
\item billboard = FALSE
}

\strong{PolygonLayer}
\itemize{
\item lineMiterLimit = 4
\item extruded = FALSE
\item wireframe = TRUE
\item elevationScale = 1
\item lineJointRounded = FALSE
\item lineWidthUnits = "pixels"
\item lineWidthScale = 1
\item stroked = TRUE
\item filled = TRUE
\item lineWidthMinPixels = 0
\item lineWidthMaxPixels = 15
}

\strong{All layers}
\itemize{
\item beforeId = NULL
\item zIndex = 1
\item autoHighlight = FALSE - requires \code{pickable} to be TRUE
\item highlightColor = c(0, 0, 128, 128) - only applicable if \code{autoHighlight} is TRUE
}

\code{zIndex} can be used to set layers order if multiple layers are added to a map.
Higher values will be plotted on top of lower values.
It is ignored, if \code{beforeId} is supplied.
}
\examples{
# default settings
renderOptions()

# modify selected options
renderOptions(radiusUnits = "meters", radiusScale = 10)


}
