% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGeoArrowS2Layer.R
\name{addGeoArrowS2Layer}
\alias{addGeoArrowS2Layer}
\title{Add Deck.gl S2Layer to a \code{\link[mapgl:maplibre]{mapgl::maplibre()}} or \code{\link[mapgl:mapboxgl]{mapgl::mapboxgl()}} map
using blazing fast \code{\link[nanoarrow:read_nanoarrow]{nanoarrow::write_nanoarrow()}} data transfer.}
\usage{
addGeoArrowS2Layer(
  map,
  data,
  source,
  file,
  url,
  layer_id = "s2",
  s2_column_name = "s2_cells",
  popup = NULL,
  tooltip = NULL,
  render_options = renderOptions(),
  data_accessors = dataAccessors(),
  popup_options = popupOptions(),
  tooltip_options = tooltipOptions(),
  ...
)
}
\arguments{
\item{map}{the \code{\link[mapgl:maplibre]{mapgl::maplibre()}} or \code{\link[mapgl:mapboxgl]{mapgl::mapboxgl()}} map to add the layer to.}

\item{data}{a \code{sf}, \code{wk}, \code{geos} or \code{SpatVector} \verb{(MULTI)POLYGON} object.
Ignored if \code{source} is supplied.}

\item{source}{the \code{id} of a source previously added via \code{\link[=addSource]{addSource()}}.}

\item{file}{a valid local file path to a \code{geoarrow} or \code{geoparquet} file to be
added to the map. Ignored if \code{source} or \code{data} is supplied.}

\item{url}{a URL to a remotely hosted \code{geoarrow} or \code{geoparquet} file to be
added to the map. Ignored if \code{source} or \code{data} or \code{file} is supplied.}

\item{layer_id}{the layer id.}

\item{s2_column_name}{the name of the S2 cells column of the data object.}

\item{popup}{should a popup be contructed? If \code{TRUE}, will create a popup fromm all
available attributes of the feature. Can also be a character vector of column
names, on which case the popup will include only those columns. If a single character
is supplied, then this will be shown for all features. If \code{NULL} (deafult) or
\code{FALSE}, no popup will be shown.}

\item{tooltip}{should a tooltip be contructed? If \code{TRUE}, will create a tooltip fromm all
available attributes of the feature. Can also be a character vector of column
names, on which case the tooltip will include only those columns. If a single character
is supplied, then this will be shown for all features. If \code{NULL} (deafult) or
\code{FALSE}, no tooltip will be shown.}

\item{render_options}{a list of \link{renderOptions}}

\item{data_accessors}{a list of \link{dataAccessors}}

\item{popup_options}{a list of \link{popupOptions}}

\item{tooltip_options}{a list of \link{tooltipOptions}}

\item{...}{can be used to pass additional props and parameters to the deck.gl
instance. See Details for more info.}
}
\value{
The modified \code{map} object with the added polygon layer.
}
\description{
Add Deck.gl S2Layer to a \code{\link[mapgl:maplibre]{mapgl::maplibre()}} or \code{\link[mapgl:mapboxgl]{mapgl::mapboxgl()}} map
using blazing fast \code{\link[nanoarrow:read_nanoarrow]{nanoarrow::write_nanoarrow()}} data transfer.
}
\details{
\code{...} can be used to pass additional props and parameters to the deck.gl instance
for fine-tuning rendering behaviour. For example, we can pass a list called
\code{parameters} with settings that control the GPU pipeline of the deck.gl instance.
See \url{https://luma.gl/docs/api-reference/core/parameters} for a list of
available prarmeters.

By default, all deck.gl layers passed to a \code{maplibre()} map will be drawn on
top of existing ones. It is, however, possible to inject layers into the
existing \code{maplibre} (base) layer stack by using
\code{render_options = renderOptions(beforeId = "<some-existing-layer-id>")}
which will plot the current layer underneath \code{"<some-existing-layer-id>"}.
See below for an example.
}
\examples{
library(wk)
library(s2)
library(mapgl)

## global coverage at S2 level 3
n = 1e5

pts = data.frame(
  id = seq_len(n)
  , geometry = xy(
    x = runif(n, -180, 180)
    , y = runif(n, -90, 90)
    , crs = 4326
  )
)

pts$s2cell = as_s2_cell(pts$geometry)
pts$s2parent_l4 = as.character(s2_cell_parent(pts$s2cell, level = 3))
pts$s2cell = as.character(pts$s2cell)

sbs = pts[!duplicated(pts$s2parent_l4), ]

style_positron = "https://basemaps.cartocdn.com/gl/positron-gl-style/style.json"

m = maplibre(style = style_positron)

m |>
  addGeoArrowS2Layer(
    data = sbs
    , layer_id = "s2layer"
    , s2_column_name = "s2parent_l4"
    , data_accessors = dataAccessors(
      getFillColor = "#74aa2380"
      , getLineColor = "#4523bb"
      , getLineWidth = 2
    )
  ) |>
  set_view(c(0, 0), 2) |>
  add_globe_control() |>
  add_navigation_control(visualize_pitch = TRUE)

## S2 hierarchy for one point
dat = data.frame(
  token = as.character(
    s2_cell_parent(
      as_s2_cell(
        s2_lnglat(-75.7019612, 45.4186427)
      )
      , level = 1:30
    )
  )
  , elevation = seq(10, 1000, length.out = 30)
  , fillColor = hcl.colors(30, palette = "inferno", alpha = 0.1)
)

m = maplibre()

m |>
  addGeoArrowS2Layer(
    data = dat
    , layer_id = "s2layer"
    , s2_column_name = "token"
    , render_options = renderOptions(
      extruded = TRUE
      , wireframe = TRUE
    )
    , data_accessors = dataAccessors(
      getFillColor = "fillColor"
      , getElevation = "elevation"
    )
  ) |>
  set_view(c(-45, 45), 1) |>
  add_globe_control() |>
  add_navigation_control(visualize_pitch = TRUE)

}
