% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{renderOptions}
\alias{renderOptions}
\title{Deck.gl render options}
\usage{
renderOptions(...)
}
\arguments{
\item{...}{named options to be passed to the relevant deck.gl JavaScript Method.}
}
\value{
list with named options, possibly modified via \code{...} argument.
}
\description{
In deck.gl every layer type has a specific set of render options, see e.g.
those for \href{https://deck.gl/docs/api-reference/layers/scatterplot-layer#render-options}{ScatterPlotLayer}.
This function sets all defaults for all available layer functions in this
package.
Please refer to the relevant \href{https://deck.gl/docs/api-reference/layers/}{deck.gl documentation}
for a more detailed description of the available layer functions.
See Details for a list of currently available options, their defaults and the
layer types they apply to.
}
\details{
Currently, the following options are automatically set to he following defaults:
\itemize{
\item radiusUnits = "pixels" (ScatterplotLayer)
\item radiusScale = 1 (ScatterplotLayer)
\item lineWidthUnits = "pixels" (ScatterplotLayer, PolygonLayer)
\item lineWidthScale = 1 (ScatterplotLayer, PolygonLayer)
\item stroked = TRUE (ScatterplotLayer, PolygonLayer)
\item filled = TRUE (ScatterplotLayer, PolygonLayer)
\item radiusMinPixels = 3 (ScatterplotLayer)
\item radiusMaxPixels = 15 (ScatterplotLayer)
\item lineWidthMinPixels = 0 (ScatterplotLayer, PolygonLayer)
\item lineWidthMaxPixels = 15 (ScatterplotLayer, PolygonLayer)
\item billboard = FALSE (ScatterplotLayer, PathLayer)
\item antialiasing = FALSE (ScatterplotLayer)
\item extruded = FALSE (PolygonLayer)
\item wireframe = TRUE (PolygonLayer)
\item elevationScale = 1 (PolygonLayer)
\item lineJointRounded = FALSE (PolygonLayer)
\item lineMiterLimit = 4 (PolygonLayer)
\item widthUnits = "pixels" (PathLayer)
\item widthScale = 1 (PathLayer)
\item widthMinPixels = 1 (PathLayer)
\item widthMaxPixels = 5 (PathLayer)
\item capRounded = TRUE (PathLayer)
\item jointRounded = FALSE (PathLayer)
\item miterLimit = 4 (PathLayer)
\item beforeId = NULL (all Layer types)
}
}
\examples{
# default settings
renderOptions()

# modify selected options
renderOptions(radiusUnits = "meters", radiusScale = 10)


}
